+++ public/js/admin.master.skill.js
@@ -72,6 +72,8 @@ __webpack_require__.r(__webpack_exports__);
       var vm = this;
       vm.showList = true;
       vm.showForm = false;
+      vm.isEdit = false;
+      vm.thisUuid = '';
     },
     setShowForm: function setShowForm() {
       var vm = this;
@@ -119,7 +121,7 @@ __webpack_require__.r(__webpack_exports__);
     addDetail: function addDetail(element) {
       var app = this;
       var el = $(element).parents('tr');
-      var table = "<tr class='details' style='background: #FFFFFF'><td colspan='5' align='center'><table width='100%'><tr><td>test</td></tr></table></td></tr>";
+      var table = "<tr class='details' style='background: #FFFFFF'><td colspan='5' align='center'><table width='100%'><tr><td>" + app.dataDetail.description + "</td></tr></table></td></tr>";
 
       if (!el.next().hasClass('details')) {
         if ($('tr').hasClass('details')) {
@@ -139,14 +141,28 @@ __webpack_require__.r(__webpack_exports__);
     },
     deleteData: function deleteData(uuid) {
       var vm = this;
-      vm.$http({
-        url: "".concat(vm.apiUrl, "/skill/").concat(uuid, "/delete"),
-        method: 'DELETE'
-      }).then(function (res) {
-        vm.$refs.table.reload();
-        toastr.success(res.data.message, 'Success');
-      })["catch"](function (err) {
-        toastr.error(err.response.data.message, 'Error');
+      swal({
+        title: "Apakah anda yakin?",
+        text: "Data yang dihapus tidak dapat dikembalikan.",
+        type: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#DD6B55",
+        confirmButtonText: "Yes!",
+        cancelButtonText: "No",
+        closeOnConfirm: false,
+        closeOnCancel: false
+      }).then(function (isConfirm) {
+        if (isConfirm) {
+          vm.$http({
+            url: "".concat(vm.apiUrl, "/skill/").concat(uuid, "/delete"),
+            method: 'DELETE'
+          }).then(function (res) {
+            vm.$refs.table.reload();
+            toastr.success(res.data.message, 'Success');
+          })["catch"](function (err) {
+            toastr.error(err.response.data.message, 'Error');
+          });
+        }
       });
     }
   },
@@ -231,6 +247,10 @@ __webpack_require__.r(__webpack_exports__);
 
         oFReader.onload = function (oFREvent) {
           $('.images').css('background-image', 'url(' + oFREvent.target.result + ')');
+          $('.images').css({
+            'background-size': 'cover',
+            'background-position': 'center center'
+          });
         };
       }
     },
@@ -242,6 +262,7 @@ __webpack_require__.r(__webpack_exports__);
         urls = "".concat(vm.apiUrl, "/skill/").concat(uuid, "/update");
       }
 
+      Aropex.btnLoad('.btn-submit', true);
       var formData = new FormData($("#FormTambah")[0]);
       vm.axios.post(urls, formData, {
         headers: {
@@ -249,10 +270,10 @@ __webpack_require__.r(__webpack_exports__);
         }
       }).then(function (res) {
         vm.$parent.setShowList();
-        $('.btn-submit').prop('disabled', false);
+        Aropex.btnLoad('.btn-submit', false);
         toastr.success(res.data.message, 'Success');
       })["catch"](function (err) {
-        $('.btn-submit').prop('disabled', false);
+        Aropex.btnLoad('.btn-submit', false);
         toastr.error(err.response.data.message, 'Error');
       });
     },
@@ -313,8 +334,7 @@ var render = function() {
                       on: {
                         click: function($event) {
                           $event.preventDefault()
-                          _vm.showList = false
-                          _vm.showForm = true
+                          return _vm.setShowForm()
                         }
                       }
                     },
@@ -362,8 +382,7 @@ var render = function() {
                       on: {
                         click: function($event) {
                           $event.preventDefault()
-                          _vm.showList = true
-                          _vm.showForm = false
+                          return _vm.setShowList()
                         }
                       }
                     },
@@ -439,22 +458,31 @@ var render = function() {
                 "div",
                 { staticClass: "form-group", attrs: { align: "center" } },
                 [
-                  _c("div", { staticClass: "image-upload-box images" }, [
-                    _c("input", {
-                      staticClass: "form-control",
-                      attrs: {
-                        type: "file",
-                        id: "image",
-                        accept: "image/png, image/jpeg",
-                        name: "image",
-                        required: "",
-                        placeholder: "Name"
-                      },
-                      on: { change: _vm.changeImage }
-                    }),
-                    _vm._v(" "),
-                    _vm._m(0)
-                  ])
+                  _c(
+                    "div",
+                    {
+                      staticClass: "image-upload-box images",
+                      style:
+                        "background-image: url(" +
+                        _vm.formData.image +
+                        "); background-size: cover; background-position: center center;"
+                    },
+                    [
+                      _c("input", {
+                        staticClass: "form-control",
+                        attrs: {
+                          type: "file",
+                          id: "image",
+                          accept: "image/png, image/jpeg",
+                          name: "image",
+                          placeholder: "Name"
+                        },
+                        on: { change: _vm.changeImage }
+                      }),
+                      _vm._v(" "),
+                      _vm._m(0)
+                    ]
+                  )
                 ]
               )
             ]),
@@ -510,7 +538,7 @@ var render = function() {
                     type: "text",
                     name: "link",
                     required: "",
-                    placeholder: "Name"
+                    placeholder: "Link"
                   },
                   domProps: { value: _vm.formData.link },
                   on: {
